Have bben fixed bugs in the tests

+++ os-server/src/main/java/app/openschool/coursemanagement/repository/DifficultyRepository.java
@@ -3,5 +3,5 @@ package app.openschool.coursemanagement.repository;
 import app.openschool.coursemanagement.entity.Difficulty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DifficultyRepository extends JpaRepository<Difficulty, Long> {
+public interface DifficultyRepository extends JpaRepository<Difficulty, Integer> {
 }
